update base class and extend and reuse on registration flow
